+++ src/components/icons/chevron-left.js
