+++ src/components/icons/logo.js
