+++ src/components/icons/logo_green.js
