+++ src/components/icons/magnifier.js
