+++ src/components/icons/menu_bars.js
