+++ src/components/icons/menu_close.js
