+++ src/components/icons/timer.js
