+++ src/components/icons/vegetarian_mark.js
